JCLOUDS-1520: Reuse SSLSocketFactory in JavaUrlHttpCommandExecutorService
Changed the untrusted SSLContext supplier in JavaUrlHttpCommandExecutorService to a
supplier of SSLSocketFactory to facilitate HTTPS socket reuse when jclouds.trust-all-certs
is enabled.

Classes extending JavaUrlHttpCommandExecutorService have been changed to match their
constructor signature with JavaUrlHttpCommandExecutorService's.

Existing users of the untrusted Supplier<SSLContext> have been changed to accept
a Supplier<SSLSocketFactory>.

LazyCachingSSLSocketFactorySupplier is used to retain the existing behavior in
which the SSLSocketFactory is created only when the first request is initiated and
not during the Guice Module's initialization.